refactored the project tree, the migration folder
