+++ scripts/migration/mainnet/rate-providers/deploy-st-rate.js
