+++ scripts/migration/mainnet/vaults/deploy-cb.js
@@ -2,8 +2,8 @@
  * Coinbase Ethereum LST
  */
 
-const { addresses } = require("./config-addresses");
-const { deployVault } = require("../deploy-vault");
+const { addresses } = require("../config-addresses");
+const { deployVault } = require("../../deploy-vault");
 
 const VaultName = "IncbEthVault",
   TokenName = "Inception Restaked cbETH",
