+++ scripts/migration/mainnet/vaults/deploy-ls.js
@@ -2,8 +2,8 @@
  * Liquid Staked Ethereum LST
  */
 
-const { addresses } = require("./config-addresses");
-const { deployVault } = require("../deploy-vault");
+const { addresses } = require("../config-addresses");
+const { deployVault } = require("../../deploy-vault");
 
 const VaultName = "InlsEthVault",
   TokenName = "Inception Restaked lsETH",
