+++ scripts/migration/mainnet/vaults/deploy-o.js
@@ -2,8 +2,8 @@
  * Origin Protol Ethereum LST
  */
 
-const { addresses } = require("./config-addresses");
-const { deployVault } = require("../deploy-vault");
+const { addresses } = require("../config-addresses");
+const { deployVault } = require("../../deploy-vault");
 
 const VaultName = "InoEthVault",
   TokenName = "Inception Restaked oETH",
