+++ scripts/migration/mainnet/vaults/deploy-os.js
@@ -2,8 +2,8 @@
  * StakeWise Ethereum LST
  */
 
-const { addresses } = require("./config-addresses");
-const { deployVault } = require("../deploy-vault");
+const { addresses } = require("../config-addresses");
+const { deployVault } = require("../../deploy-vault");
 
 const VaultName = "InosEthVault",
   TokenName = "Inception Restaked osETH",
