+++ scripts/migration/mainnet/vaults/deploy-st.js
@@ -1,0 +1,17 @@
+const { addresses } = require("../config-addresses");
+const { deployVault } = require("../../deploy-vault");
+
+const VaultName = "InstEthVault",
+  TokenName = "Inception Restaked stETH",
+  TokenSymbol = "InstETH";
+
+async function main() {
+  await deployVault(addresses, VaultName, TokenName, TokenSymbol);
+}
+
+main()
+  .then(() => process.exit(0))
+  .catch((error) => {
+    console.error(error);
+    process.exit(1);
+  });
